Add iterable-attributes test.
Commented (in cahinAsync) call to consume() to force "when" to start consuming. This is a result of making all the async helper functions lazy, which means a "when" (or any other helped function) that is never actually assigned might not run.

This is DESIRABLE, but not necessarily compatible, so this branch is parked and the other fixes will be back-ported to v0.10.11