Update AsyncExtaHelper types DNR / WIP
Note: this required some @ts-ignores (which need investigation) and the use of of AsyncExtraIterable when wrapping all the helpers within their own declarations (rathe than via wrapAsyncExtra) is ignoring the difference between an AsyncIterable|Iterator, which also needs to be confirmed.

OTOH, the types now work as intended in z2m-lite and the tests pass (new tests for redeclared iterables, and attempts to redeclare base attrs as iterable need implementing).