Improve resolution of ingerited iterable properties.
In thre previous implementation, we relied on simple intersection when deriving the type of an extended tag. This led to iterable types like:

Partial<AsyncExtraIterable<BaseThings>> & Partial<AsyncExtraIterable<ExtThings>>

...which in turn leads to failure to resolve the iterated type as the actual type of [Symbol.asyncIterator] was `(()=>AsyncIterator<BaseThings>) & (()=>AsyncIterator<ExtThings>)`. This represents an overloaded function for which the ReturnType is always the first declaration - subsequent ones are ignored (ie, in the above case, the iterated types appears to be "BaseThings", not "BaseThings & ExtThings".

The last impl attempted to resilve this via a recursive type that extracted each part of the intersection, but it is very slow and error-prone (blowing up the compiler, not in terms of type soundness).

In this implementation we avoid the whole problem by omitting iterable properties from the Base in the Ext, intersecting them and then re-applying IterableProperties, so they are generated as AsyncExtraIterable<BaseThung & ExtThing>.

This mechanism is much better, but imperfect as it assumes properties in the base which have the same keys as those in the Ext type are also iterable. In fact this need not be the case. A better implementation would be the BaseCreator to carry the definitions rather than just the combined effective type, so base and ext properties can be combined in different ways. For example, iterables are intersected and then wrapped in AsyncExtra<...>, whereas overrides, declares and ids can have their own type-safe derivations such as partial type-equality (overrides), absence (declares, ids)

However this impl as committed pass all the tests and runs z2m-lite correctly without error. Additional tests are required