Smarter "disconnect consuming node" logic
Async iterated attributes & nodes now process until a node is _removed_ from the DOM. Previously, they only iterated if a node was _present_ in the DOM, meaning an iterator that yielded during construction would cause termination of the conumer if the constructor itself later yielded control before the node was eventually mounted.

The current algorithm terminates consumption if the node is not present in the DOM, but previously has been.

Note: this has the potential to leak if the node is mounted and removed and the consumer doesn't yield in the mounted state, as the node will appear to have been never mounted.

The only resolution to this is another mutation observer that tracks such mount/dismount pairs, or itself handles the termination of the consumers (ie a Map<node,fnStopConsumption> which is tested for all removed nodes). Whether this is a necessary overhead depends on if the leaks are significant (TDB)