Re-implement iterable = iterator assignments
The previous implementation of iterable assignments warned but did not have a sensible implementation: the initial iterator would be left in place and the replacement left unyielding.

This has been updated so that assigning an iterator to an iterable object property now terminates the existing value (if any), and replaces it with the new assignment as and when it yields.

* Fix: assigning an iterator to an iterable no longer errorneously pushes an `undefined` to the consumer (iterators.ts:276)
* Update lib to ES2022 to allow for error.cause
* Add iter-iter test