Change nodes(...):Array<Node> to nodes(...): IterableIterator<Node>
The removes the need to reference the returned value in a closure, in case is it updated. Instead, it is lazily walked on demand, including any updates. This appears to work fine (needs full tests).

On examination, the remaining leak is when a node is removed and there is no subsequent iteration to unwind the references. This can happen, for example, when a node depends on a child it is expecting to receieve events, but the child is removed and so the container (with an implicit reference to teh child) never receives another input event from the removed child, leaving it in the async iteration state forever, having never called ai.return() (or throw)